requirements.txt
asgiref==3.7.2
astroid==2.15.6
attrs==23.1.0
autopep8==2.0.4
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
dill==0.3.7
distlib==0.3.7
dj-database-url==0.5.0
Django==3.2.20
django-magic==0.5a0
filelock==3.12.3
h11==0.14.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
libmagic==1.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pycparser==2.21
pylibmagic==0.4.0
pylint==2.17.5
PySocks==1.7.1
python-magic==0.4.27
pytz==2023.3
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomlkit==0.12.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
wrapt==1.15.0
wsproto==1.2.0
